// src/services/-products-{id}-questions-get.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Shopping Platform API
 * Production-grade e-commerce backend
 * OpenAPI spec version: 1.0
 */
export interface UtilsResponse {
  code?: number;
  error?: string;
  message?: string;
  success?: boolean;
}

export interface DtoProductAnswerResponse {
  author?: string;
  body?: string;
  created_at?: string;
  id?: number;
  is_ai_reply?: boolean;
  is_store_reply?: boolean;
  question_id?: number;
}

export interface DtoProductQuestionResponse {
  answers?: DtoProductAnswerResponse[];
  author?: string;
  body?: string;
  created_at?: string;
  id?: number;
  is_owner?: boolean;
  product_id?: number;
}

export type GetProductsIdQuestionsParams = {
/**
 * Items per page
 */
limit?: number;
/**
 * Offset
 */
offset?: number;
};

export type GetProductsIdQuestions200Data = {
  limit?: number;
  offset?: number;
  questions?: DtoProductQuestionResponse[];
  total?: number;
};

export type GetProductsIdQuestions200 = UtilsResponse & {
  data?: GetProductsIdQuestions200Data;
};